TC#47: Structured types on the backend
