+++ backend/src/types/carousel.d.ts
@@ -1,0 +1,5 @@
+export interface CarouselFields {
+  image: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
