+++ backend/src/types/category.d.ts
@@ -1,0 +1,3 @@
+export interface CategoryFields {
+  name: string;
+}
