+++ backend/src/types/footer.d.ts
@@ -1,0 +1,11 @@
+export interface LinksMediaFields {
+  name: string;
+  value: string;
+}
+
+export interface FooterFields {
+  socialNetwork: LinksMediaFields[];
+  menuPosition: LinksMediaFields[];
+  publicOffer: string;
+  mainPartnerImage: string | null;
+}
