+++ backend/src/types/news.d.ts
@@ -1,0 +1,9 @@
+export interface NewsFields {
+  title: string;
+  subtitle?: string;
+  content: string;
+  createdAt: Date;
+  updatedAt: Date;
+  newsCover?: string;
+  images?: string[];
+}
