+++ backend/src/types/partner.d.ts
@@ -1,0 +1,5 @@
+export interface PartnerFields {
+  name: string;
+  image: string;
+  url: string;
+}
