+++ backend/src/utils/utils.ts
@@ -2,18 +2,7 @@ import * as handlebars from 'handlebars';
 import * as fs from 'node:fs';
 import nodemailer from 'nodemailer';
 import path from 'path';
-import { UserFields } from '../../types';
-
-// В этом файле можно удобно хранить утилиты.
-// Например у вас будет много мест с проверкой данных приходящих с фронта
-// если вам надо проверить значение на валидность
-// Типа строка это или число или не нулл и тд
-// Можно создать соответствующие функции. Тут уже на ваше усмотрение.
-
-// Вот эту функцию можно переделать для более универсального использования.
-// Вообще если у вас больше ни для чего не используется отправка емелов, то тогда не надо
-// А если будет еще где-то использоваться в нескольких местах, то переделать и аргументами передавать нужные пропсы
-// Пока можно и так оставить
+import { UserFields } from '../types/user';
 
 export const sendMail = async (user: UserFields, token: string) => {
   const __dirname = path.resolve();
