+++ tests/INSTRUCTIONS.md
@@ -1,11 +1,11 @@
-# Инструкции для запуска тестов
+Инструкция для запуска тестового окружения
 
-## 1. Запустите фикстуры
-npm run seed
-
-## 2. Запустите backend и fronted
-npm run dev
-
-## 3. Запустите тесты с помощью команды:
-npx codeceptjs run --steps
-
+1. Убедитесь, что в вашем бэкенде есть файлы .test.env и .env
+2. На вашем фронтенде должен быть файл .env.local
+3. Перед первым запуском перейдите в директорию tests:
+   дайте доступ скрипту tests.sh: делаете chmod +x tests.sh
+   установите глобально пакет pm2: npm i -g pm2
+4. Все вы можете запустить скрипт с тестами и тестовым окружением прописав в терминале ./tests.sh, находясь в директории
+   tests
+5. Вы можете запустить(желательно так и делать) только определенные тесты задав определенному сценарию тег(пример можете
+   посмотреть в users.feature). Ну и запускаете, например ./tests.sh --grep @login.
